version 0.1.3.1

+++ src/main/java/Gestionnaire.java
@@ -5,6 +5,8 @@ import java.util.Deque;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
+import java.util.concurrent.CompletableFuture;
+import java.util.concurrent.ConcurrentLinkedDeque;
 import java.util.stream.Collectors;
 
 import com.sun.tools.javac.util.Pair;
@@ -14,11 +16,11 @@ import com.sun.tools.javac.util.Pair;
  * Il contient une liste de téléchargements en attente
  */
 public class Gestionnaire {
-	// Liste des telechargements
-	private final Deque<Launcher> newQueue = new ArrayDeque<>();    // La file d'attente des téléchagements instanciés
-	private final Deque<Launcher> waitQueue = new ArrayDeque<>();   // La file d'attente des téléchagements en pause
-	private final Deque<Launcher> launchQueue = new ArrayDeque<>(); // La file d'attente des téléchagements en cours de téléchargements
-	private final Deque<Launcher> endQueue = new ArrayDeque<>();    // La file d'attente des téléchagements finis (ou interrompus)
+	// Liste des telechargements (a synchro)
+	private final Deque<Launcher> newQueue = new ConcurrentLinkedDeque<>();    // La file d'attente des téléchagements instanciés
+	private final Deque<Launcher> waitQueue = new ConcurrentLinkedDeque<>();   // La file d'attente des téléchagements en pause
+	private final Deque<Launcher> launchQueue = new ConcurrentLinkedDeque<>(); // La file d'attente des téléchagements en cours de téléchargements
+	private final Deque<Launcher> endQueue = new ConcurrentLinkedDeque<>();    // La file d'attente des téléchagements finis (ou interrompus)
 	
 	/**
 	 * Dernier launcher non lancé
@@ -45,12 +47,25 @@ public class Gestionnaire {
 		
 	}
 	
+	public boolean changeCurrentLauncher(String nom,Deque<Launcher> queue) {
+		Launcher l=queue.parallelStream().reduce(null, (a,e) -> e.getNom().equals(nom)?e:null);
+		if (l != null) {
+			queue.remove(l);
+			queue.push(l);
+			return true;
+		}
+		return false;
+	}
+	
 	// Lance le launcher au dessus de la pile
 	public boolean launch() {
 		if (getCurrentNew() != null) {
+			if(this.getCurrentNew().getEtat() != Launcher.state.NEW) {
+				return false;
+			}
 			Launcher currentNew = newQueue.pop();
-			currentNew.start();
 			launchQueue.push(currentNew);
+			currentNew.start().thenAccept(e -> { if(launchQueue.remove(currentNew)) { endQueue.add(currentNew); }});			
 			return true;
 		}
 		return false;
@@ -62,34 +77,77 @@ public class Gestionnaire {
 	 * @return : si celui ci n'existe pas renvoie faux
 	 */
 	public boolean launch(String launcher) {
-		if (!changeNewCurrentLauncher(launcher,newQueue)) {
+		if (!changeCurrentLauncher(launcher,newQueue)) {
 			return false;
 		}
-		this.launch();
-		return true;
+		return this.launch();
 		
 	}
+	
+	public boolean delete() {
+		if (getCurrentLaunch() != null) {
+			Launcher l = launchQueue.pop();
+			l.delete();
+			endQueue.add(l);
+			return true;
+		}
+		return false;
+	}
+	
+	public boolean delete(String launcher) {
+		if (!changeCurrentLauncher(launcher,launchQueue)) {
+			return false;
+		}
+		return this.delete();
+		
+	}
+	
+	public boolean pause() {
+		if (getCurrentLaunch() != null) {
+			if(this.getCurrentLaunch().getEtat() != Launcher.state.WORK) {
+				return false;
+			}
+			Launcher l = launchQueue.pop();
+			l.pause();
+			waitQueue.add(l);
+			return true;
+		}
+		return false;
+	}
+	
+	public boolean pause(String launcher) {
+		if (!changeCurrentLauncher(launcher,launchQueue)) {
+			return false;
+		}
+		return this.pause();
+	}
+	
+	public boolean restart() {
+		if (getCurrentWait() != null) {
+			if(this.getCurrentWait().getEtat() != Launcher.state.WAIT) {
+				return false;
+			}
+			Launcher l = waitQueue.pop();
+			launchQueue.push(l);
+			l.restart().thenAccept((e) -> { if(launchQueue.remove(l)) endQueue.add(l); });
+		}
+		return false;
+	}
+	
+	public boolean restart(String launcher) {
+		if (!changeCurrentLauncher(launcher,waitQueue)) {
+			return false;
+		}
+		return this.restart();
+	}
+	
 	
 	public void addLauncher(String URL) {
 		Launcher l = new LauncherTelechargement(URL);
 		newQueue.push(l);
 	}
 	
-	public void delete() {
-		Launcher l = launchQueue.pop();
-		l.delete();
-		endQueue.add(l);
-	}
-	
-	public boolean changeNewCurrentLauncher(String nom,Deque<Launcher> queue) {
-		Launcher l=queue.parallelStream().reduce(null, (a,e) -> e.getNom().equals(nom)?e:null);
-		if (l != null) {
-			queue.remove(l);
-			queue.push(l);
-			return true;
-		}
-		return false;
-	}
+
 	
 	// Liste des noms et etats des launchers non lancé
 	public List<Pair<String, Launcher.state>> listNew() {
